# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SheetsExportRequest": {
        "properties": {
          "month": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "sheet": {
            "title": "Sheet",
            "type": "string"
          }
        },
        "required": [
          "sheet"
        ],
        "title": "SheetsExportRequest",
        "type": "object"
      },
      "SheetsExportResult": {
        "properties": {
          "rows": {
            "title": "Rows",
            "type": "integer"
          }
        },
        "required": [
          "rows"
        ],
        "title": "SheetsExportResult",
        "type": "object"
      },
      "SheetsStatus": {
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "sa_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sa Email"
          }
        },
        "required": [
          "configured",
          "sa_email"
        ],
        "title": "SheetsStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Poi Poi Hisab API",
    "version": "0.28.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/export/sheets": {
      "post": {
        "operationId": "export_sheets_api_v1_export_sheets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SheetsExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SheetsExportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Sheets",
        "tags": [
          "export"
        ]
      }
    },
    "/api/v1/export/sheets/status": {
      "get": {
        "operationId": "sheets_status_api_v1_export_sheets_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SheetsStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sheets Status",
        "tags": [
          "export"
        ]
      }
    },
...
  }
}